third_party/blink/web_tests/external/wpt/css/css-text/animations/text-indent-interpolation-expected.txt

This is a testharness.js-based test.
Found 98 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (-0.3) should be [-15px hanging]
  assert_equals: expected "- 15px hanging " but got "- 15px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0) should be [0 hanging]
  assert_equals: expected "0px hanging " but got "0px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0.3) should be [15px hanging]
  assert_equals: expected "15px hanging " but got "15px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0.6) should be [30px hanging]
  assert_equals: expected "30px hanging " but got "30px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (1) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "50px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (1.5) should be [75px hanging]
  assert_equals: expected "75px hanging " but got "75px "
[FAIL] Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (-0.3) should be [-15px hanging]
  assert_equals: expected "- 15px hanging " but got "- 15px "
[FAIL] Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0) should be [0 hanging]
  assert_equals: expected "0px hanging " but got "0px "
[FAIL] Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0.3) should be [15px hanging]
  assert_equals: expected "15px hanging " but got "15px "
[FAIL] Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0.6) should be [30px hanging]
  assert_equals: expected "30px hanging " but got "30px "
[FAIL] Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (1) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "50px "
[FAIL] Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (1.5) should be [75px hanging]
  assert_equals: expected "75px hanging " but got "75px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (-0.3) should be [-15px hanging each-line]
  assert_equals: expected "- 15px hanging each - line " but got "- 15px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0) should be [0 hanging each-line]
  assert_equals: expected "0px hanging each - line " but got "0px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.3) should be [15px hanging each-line]
  assert_equals: expected "15px hanging each - line " but got "15px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.6) should be [30px hanging each-line]
  assert_equals: expected "30px hanging each - line " but got "30px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1) should be [50px hanging each-line]
  assert_equals: expected "50px hanging each - line " but got "50px "
[FAIL] CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1.5) should be [75px hanging each-line]
  assert_equals: expected "75px hanging each - line " but got "75px "
[FAIL] Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (-0.3) should be [-15px hanging each-line]
  assert_equals: expected "- 15px hanging each - line " but got "- 15px "
[FAIL] Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0) should be [0 hanging each-line]
  assert_equals: expected "0px hanging each - line " but got "0px "
[FAIL] Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.3) should be [15px hanging each-line]
  assert_equals: expected "15px hanging each - line " but got "15px "
[FAIL] Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.6) should be [30px hanging each-line]
  assert_equals: expected "30px hanging each - line " but got "30px "
[FAIL] Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1) should be [50px hanging each-line]
  assert_equals: expected "50px hanging each - line " but got "50px "
[FAIL] Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1.5) should be [75px hanging each-line]
  assert_equals: expected "75px hanging each - line " but got "75px "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "- 15px hanging "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "0px hanging "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "15px hanging "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "25px hanging "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "30px hanging "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "75px hanging "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "- 15px hanging "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "0px hanging "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "15px hanging "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "25px hanging "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "30px hanging "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "75px hanging "
[FAIL] CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "- 15px hanging "
[FAIL] CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "0px hanging "
[FAIL] CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "15px hanging "
[FAIL] CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "25px hanging "
[FAIL] CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "30px hanging "
[FAIL] CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "75px hanging "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "- 15px hanging "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "0px hanging "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "15px hanging "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "25px hanging "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "30px hanging "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "75px hanging "
[FAIL] CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "- 15px "
[FAIL] CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "0px "
[FAIL] CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "15px "
[FAIL] CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "25px "
[FAIL] CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "30px "
[FAIL] CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (1) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "50px "
[FAIL] CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "75px "
[FAIL] Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "- 15px "
[FAIL] Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "0px "
[FAIL] Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [0px each-line]
  assert_equals: expected "0px each - line " but got "15px "
[FAIL] Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "25px "
[FAIL] Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "30px "
[FAIL] Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (1) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "50px "
[FAIL] Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging]
  assert_equals: expected "50px hanging " but got "75px "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [0px]
  assert_equals: expected "0px " but got "- 15px hanging each - line "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (0) should be [0px]
  assert_equals: expected "0px " but got "0px hanging each - line "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [0px]
  assert_equals: expected "0px " but got "15px hanging each - line "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "25px hanging each - line "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "30px hanging each - line "
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "75px hanging each - line "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [0px]
  assert_equals: expected "0px " but got "- 15px hanging each - line "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (0) should be [0px]
  assert_equals: expected "0px " but got "0px hanging each - line "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [0px]
  assert_equals: expected "0px " but got "15px hanging each - line "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "25px hanging each - line "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "30px hanging each - line "
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "75px hanging each - line "
[FAIL] CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "- 15px hanging each - line "
[FAIL] CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (0) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "0px hanging each - line "
[FAIL] CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "15px hanging each - line "
[FAIL] CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "25px hanging each - line "
[FAIL] CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "30px hanging each - line "
[FAIL] CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "75px hanging each - line "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "- 15px hanging each - line "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (0) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "0px hanging each - line "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "15px hanging each - line "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "25px hanging each - line "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "30px hanging each - line "
[FAIL] CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "75px hanging each - line "
[FAIL] CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [0px]
  assert_equals: expected "0px " but got "- 15px "
[FAIL] CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [0px]
  assert_equals: expected "0px " but got "15px "
[FAIL] CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "25px "
[FAIL] CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "30px "
[FAIL] CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (1) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "50px "
[FAIL] CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "75px "
[FAIL] Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [0px]
  assert_equals: expected "0px " but got "- 15px "
[FAIL] Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [0px]
  assert_equals: expected "0px " but got "15px "
[FAIL] Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "25px "
[FAIL] Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "30px "
[FAIL] Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (1) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "50px "
[FAIL] Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging]
  assert_equals: expected "50px hanging each - line " but got "75px "
Harness: the test ran to completion.
